+++ NEWS.md
@@ -1,3 +1,8 @@
+# ahnr 0.3.1
+
+* Removing generics.
+* plot function name changed to visualize.
+
 # ahnr 0.3.0
 
 * Generic functions updated: summary, predict, plot.